chore: slimmed down & integrated cardiconview, models folder, cleaned up file structure
