+++ stoa/stoa/Helpers/Color+Extensions.swift
@@ -1,0 +1,24 @@
+//
+//  Color+Extensions.swift
+//  stoa
+//
+//  Created by Agustio Maitimu on 20/07/25.
+//
+
+import SwiftUI
+
+extension Color {
+	init(hex: String) {
+		let scanner = Scanner(string: hex)
+		_ = scanner.scanString("#")
+		
+		var rgb: UInt64 = 0
+		scanner.scanHexInt64(&rgb)
+		
+		let r = Double((rgb >> 16) & 0xFF) / 255
+		let g = Double((rgb >> 8) & 0xFF) / 255
+		let b = Double(rgb & 0xFF) / 255
+		
+		self.init(red: r, green: g, blue: b)
+	}
+}
