+++ stoa/stoa/Models/quotes.json
