+++ stoa/stoa/View/PhilosopherQuotesView.swift
